+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(bvartools)
+
+test_check("bvartools")